Refactor antipatterns and comments
Multiline args and kwargs:

The changed function already uses 3 lines.
This change groups args and kwargs keeping the number of lines of code.

Antipatterns:

Trying to get an attribute from an object and excepting an
AttributeException if it fails can be improved by using the `getattr`
builtin method and passing a default value.

Using boolean algebra to determine the value of a variable can lead to
antipatterns.
In this case `isFalsy(value1) and value2 or value1` is oldschool for
`value2 if isFalsy(value1) else value1`, yet it can be simplified to
`value1 or value2`.

Comments

opengever.core has different kinds of docstring conventions:
    """This is a docstring convention"""
    """This is another docstring convention
    """

The first convention is standard in Plone 4.3 and 5.0:
https://docs.plone.org/4/en/external/plone.api/docs/contribute/conventions.html#docstrings-style
https://docs.plone.org/develop/styleguide/python.html#docstrings-style
Therefore all docstrings with the second convention will be changed to
the first convention.