Move fields to common tab
When a field is not in a fieldset, it appears in the "default" or
"standard" tab.
These fields belong into the "common" tab.

Also:
Put title on top of everything (and test it)

The test environment does not have always both languages, for this
reason the assertion is implemented in a custom method.