Default Values: Fix edge case for title and description fields:
When a title or description field has a default(factory), there
is one more pathological case where defaults weren't persisted
correctly up until now:

The constructor in P.CMFCore.PortalFolder.PortalFolderBase sets
title and description specificially to the empty string if no
*actual* field value was provided.

This then fools our own set_default_values() which is called
later: Because there already is a value persisted on the
object for those fields, it doesn't attempt to determine and
set default values. It has no way of knowing whether that
empty string value was the value actually intended by the user,
or the hardcoded value from the PortalFolderBase constructor.

We therefore treat these fields separately, and check if they
have defaults *before* object construction. If so, and no
*actual* value has been provided for the field in question,
we inject the determined default into the `kw` dict, so that
it then gets treated like it were a regular value.